[FE] FIX: useRoom query 파일 분리 #147
useQuery는 호출되면 자동으로 실행되기 때문에 파일을 분리해야 합니다.
