+++ src/frontend/src/pages/HomePage/index.tsx
@@ -1,7 +1,7 @@
 import { VideoGridContainer } from './index.css';
 import { VideoCard } from '@/components/VideoCard';
 import { useNavigate } from 'react-router-dom';
-import { useRoom } from '@/hooks/queries/useRoom';
+import { useRooms } from '@/hooks/queries/useRooms';
 import { useEffect, useState } from 'react';
 import { RoomDto } from '@/api/endpoints/room/room.interface';
 import { HomeSkeleton } from './HomeSkeleton';
@@ -9,7 +9,7 @@ import { useDelayedLoading } from '@/hooks/utils/useDelayedLoading';
 
 export const HomePage = () => {
   const navigate = useNavigate();
-  const { getRooms } = useRoom();
+  const getRooms = useRooms();
   const showSkeleton = useDelayedLoading(getRooms.data);
   const [videos, setVideos] = useState<RoomDto[]>([]);
